requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
torch>=1.10.0
torchaudio>=0.10.0

# Audio processing
librosa>=0.9.0
soundfile>=0.10.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0

# Configuration
pyyaml>=5.4.0

# Jupyter notebooks
jupyter>=1.0.0
ipython>=7.0.0

# Optional: Dimensionality reduction
umap-learn>=0.5.0





